% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_efa.R
\name{step_efa}
\alias{step_efa}
\title{Factor Analysis (Exploratory)}
\usage{
step_efa(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  num_comp = 2,
  res = NULL,
  prefix = "MC",
  skip = FALSE,
  id = recipes::rand_id("efa")
)
}
\arguments{
\item{...}{One or more selector functions to choose
which variables will be used to compute the efa.}

\item{role}{For model terms created by this step, what analysis
role should they be assigned?. By default, the function assumes
that the new multiple corespondence columns created by the
original variables will be used as predictors in a model}

\item{num_comp}{The number of MC variables to retain as new
predictors. Currently throws an error if non-positive values
are used.}

\item{res}{The [psych::fa()] object is stored here once
this preprocessing step has been trained by [prep.recipe()]}

\item{prefix}{A character string that will be the prefix for
the new MC variables. Defaults to "MC".}
}
\value{
An updated version of `recipe` with the new step
 added to the sequence of existing steps (if any).
}
\description{
`step_efa` creates a *specification* of a
 recipe step that will compute a factor analysis
 of a set of variables.
}
\details{
Factor analysis is a dimension reduction
method that uses estimation to rotate variables
rather than the empirical PCA method.
It seeks to find commonality between variables and
lump these into one factor (not to be confused with
the data type!). This recipe step assumes linearity
whereas step_efa_poly does not.

At this juncture, unlike the proper recipes found
within `tidymodels`, the variable names are not
padded with zeros.
}
\examples{
# Using the Farms dataset
example_data <- psych::bfi

# Original Data:
head(example_data)

example_rec <- recipes::recipe(age ~ ., data = example_data)

example_efa <- step_efa(example_rec, recipes::all_predictors(), num_comp = 3)

# Prep the recipe

efa_prepped <- recipes::prep(example_efa)

# Transformed Data:

transformed_data <- recipes::juice(efa_prepped)

head(transformed_data)



}
\references{
Goldberg, L.R. (1999) A broad-bandwidth,
 public domain, personality inventory measuring
 the lower-level facets of several five-factor models.
 In Mervielde, I. and Deary, I. and De Fruyt, F.
 and Ostendorf, F. (eds) Personality psychology in Europe. 7.
 Tilburg University Press. Tilburg, The Netherlands.
}
\concept{fa}
\concept{preprocessing}
\concept{projection_methods}
\keyword{datagen}
