% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_mca.R
\name{step_mca}
\alias{step_mca}
\title{Multiple Correspondence Analysis}
\usage{
step_mca(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  num_f = 2,
  res = NULL,
  prefix = "MC",
  skip = FALSE,
  id = recipes::rand_id("mca")
)
}
\arguments{
\item{...}{One or more selector functions to choose
which variables will be used to compute the MCA.}

\item{role}{For model terms created by this step, what analysis
role should they be assigned?. By default, the function assumes
that the new multiple corespondence columns created by the
original variables will be used as predictors in a model}

\item{num_f}{The number of MC variables to retain as new
predictors. Currently throws an error if non-positive values
are used.}

\item{res}{The [MASS::mca()] object is stored here once
this preprocessing step has been trained by [prep.recipe()]}

\item{prefix}{A character string that will be the prefix for
the new MC variables. Defaults to "MC".}
}
\value{
An updated version of `recipe` with the new step
 added to the sequence of existing steps (if any).
}
\description{
`step_mca` creates a *specification* of a
recipe step that will compute a multiple
correspondence analysis of a set of factors.
This uses the `mca` function from the `MASS`
package to acheive this.
}
\details{
Multiple Correspondence Analysis is a
 dimension reduction method that works with
 categorical variables only, addressing the
 shortcomings of other dimention-reducing methods
 such as PCA.

At this juncture, unlike the proper recipes found
within `tidymodels`, the variable names are not
padded with zeros.
}
\examples{
# Using the Farms dataset
example_data <- MASS::farms

# Original Data:
example_data

example_rec <- recipes::recipe(Mois ~ ., data = example_data)

example_mca <- step_mca(example_rec, recipes::all_predictors(), num_f = 5)

# Prep the recipe

mca_prepped <- recipes::prep(example_mca)

# Transformed Data:
transformed_data <- recipes::juice(mca_prepped)

head(transformed_data)











}
\references{
Venables, W. N. and Ripley, B. D. (2002). Modern
 Applied Statistics with S. Fourth edition. Springer.
}
\concept{mca}
\concept{preprocessing}
\concept{projection_methods}
\keyword{datagen}
